Relocate ArrayMultipleChoiceField
