+++ dive_content/forms.py
@@ -1,20 +1,8 @@
 from django import forms
 
 from .choices import *
+from .fields import ArrayMultipleChoiceField
 from .models import Leaf, Plant
-
-
-class ArrayMultipleChoiceField(forms.MultipleChoiceField):
-    def __init__(self, model=None, field_name="", **kwargs):
-        if model and field_name:
-            _label = model._meta.get_field(field_name).base_field.verbose_name
-        else:
-            _label = None
-
-        required = kwargs.pop("required", False)
-        label = kwargs.pop("label", _label)
-
-        super().__init__(required=required, label=label, **kwargs)
 
 
 class PlantAdminForm(forms.ModelForm):
